% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func2.R
\name{func2}
\alias{func2}
\title{A Cat Function}
\usage{
func2(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
func2()
}
\keyword{cats}
