% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbubble.R
\name{plotdomains}
\alias{plotdomains}
\title{Plot the Pfam Domains}
\usage{
plotdomains(maf.file, pfam.file)
}
\arguments{
\item{maf.file}{The .maf containing mutations for one gene}

\item{pfam.file}{The df object of the graphic json from jsonlite}
}
\description{
deal with graphic .json file scraped from pfam and imported using jsonlite
}
\details{
useful info : pfam$length , pfam$regions , pfam$motifs, 
regions: text, start, end , motifs:start end
start, end should be converted to numeric
}
\examples{
plotdomains(p53.maf,pfam.maf)
}
\seealso{
\code{\link{}}
}
